don't repeat yourself
