+++ bgmi/script.py
@@ -1,61 +1,17 @@
 # coding=utf-8
 from __future__ import print_function, unicode_literals
 
-import re
-import os
-import imp
 import glob
+import imp
+import os
 import time
 import traceback
 
-from bgmi.models import STATUS_UPDATED, STATUS_FOLLOWED
-from bgmi.utils import print_success, print_warning, print_info
 from bgmi.config import SCRIPT_PATH
 from bgmi.download import download_prepare
+from bgmi.models import STATUS_UPDATED, STATUS_FOLLOWED
 from bgmi.models import Script
-
-
-FETCH_EPISODE_ZH = re.compile("第?\s?(\d{1,3})\s?[話话集]")
-FETCH_EPISODE_WITH_BRACKETS = re.compile('[【\[](\d+)\s?(?:END)?[】\]]')
-FETCH_EPISODE_ONLY_NUM = re.compile('^([\d]{2,})$')
-FETCH_EPISODE_RANGE = re.compile('[\d]{2,}\s?-\s?([\d]{2,})')
-FETCH_EPISODE_OVA_OAD = re.compile('([\d]{2,})\s?\((?:OVA|OAD)\)]')
-FETCH_EPISODE_WITH_VERSION = re.compile('[【\[](\d+)\s? *v\d(?:END)?[】\]]')
-FETCH_EPISODE = (
-    FETCH_EPISODE_ZH, FETCH_EPISODE_WITH_BRACKETS, FETCH_EPISODE_ONLY_NUM,
-    FETCH_EPISODE_RANGE,
-    FETCH_EPISODE_OVA_OAD, FETCH_EPISODE_WITH_VERSION)
-
-
-def parse_episode(episode_title):
-    """
-    parse episode from title
-
-    :param episode_title: episode title
-    :type episode_title: str
-    :return: episode of this title
-    :rtype: int
-    """
-
-    _ = FETCH_EPISODE_ZH.findall(episode_title)
-    if _ and _[0].isdigit():
-        return int(_[0])
-
-    _ = FETCH_EPISODE_WITH_BRACKETS.findall(episode_title)
-    if _ and _[0].isdigit():
-        return int(_[0])
-
-    _ = FETCH_EPISODE_WITH_VERSION.findall(episode_title)
-    if _ and _[0].isdigit():
-        return int(_[0])
-
-    for split_token in ['【', '[', ' ']:
-        for i in episode_title.split(split_token):
-            for regexp in FETCH_EPISODE:
-                match = regexp.findall(i)
-                if match and match[0].isdigit():
-                    return int(match[0])
-    return 0
+from bgmi.utils import print_success, print_warning, print_info
 
 
 class ScriptRunner(object):
