+++ bgmi/website/base.py
@@ -8,11 +8,12 @@ import string
 import time
 from collections import defaultdict
 from itertools import chain
+
 import bgmi.config
 from bgmi.config import MAX_PAGE
 from bgmi.models import Bangumi, Filter, Subtitle, STATUS_FOLLOWED, STATUS_UPDATED
 from bgmi.script import ScriptRunner
-from bgmi.utils import (print_warning, print_info,
+from bgmi.utils import (parse_episode, print_warning, print_info,
                         test_connection, get_terminal_col, GREEN, YELLOW, COLOR_END)
 
 if bgmi.config.IS_PYTHON3:
@@ -21,51 +22,8 @@ else:
     _unicode = unicode
 
 
-FETCH_EPISODE_ZH = re.compile("第?\s?(\d{1,3})\s?[話话集]")
-FETCH_EPISODE_WITH_BRACKETS = re.compile('[【\[](\d+)\s?(?:END)?[】\]]')
-FETCH_EPISODE_ONLY_NUM = re.compile('^([\d]{2,})$')
-FETCH_EPISODE_RANGE = re.compile('[\d]{2,}\s?-\s?([\d]{2,})')
-FETCH_EPISODE_OVA_OAD = re.compile('([\d]{2,})\s?\((?:OVA|OAD)\)]')
-FETCH_EPISODE_WITH_VERSION = re.compile('[【\[](\d+)\s? *v\d(?:END)?[】\]]')
-FETCH_EPISODE = (
-    FETCH_EPISODE_ZH, FETCH_EPISODE_WITH_BRACKETS, FETCH_EPISODE_ONLY_NUM,
-    FETCH_EPISODE_RANGE,
-    FETCH_EPISODE_OVA_OAD, FETCH_EPISODE_WITH_VERSION)
-
-
 class BaseWebsite(object):
-    # todo : 00-11这种识别错误
-    @staticmethod
-    def parse_episode(episode_title):
-        """
-        parse episode from title
-
-        :param episode_title: episode title
-        :type episode_title: str
-        :return: episode of this title
-        :rtype: int
-        """
-
-        _ = FETCH_EPISODE_ZH.findall(episode_title)
-        if _ and _[0].isdigit():
-            return int(_[0])
-
-        _ = FETCH_EPISODE_WITH_BRACKETS.findall(episode_title)
-        if _ and _[0].isdigit():
-            return int(_[0])
-
-        _ = FETCH_EPISODE_WITH_VERSION.findall(episode_title)
-        if _ and _[0].isdigit():
-            return int(_[0])
-
-        for split_token in ['【', '[', ' ']:
-            for i in episode_title.split(split_token):
-                for regexp in FETCH_EPISODE:
-                    match = regexp.findall(i)
-                    if match and match[0].isdigit():
-                        return int(match[0])
-        return 0
-
+    parse_episode = parse_episode
     def search(self, keyword='', count=1, filter_=None):
         if not filter_:
             filter_ = '(.*)'
