+++ script_example.py
@@ -1,15 +1,15 @@
 # coding=utf-8
 from __future__ import print_function, unicode_literals
 
-import re
 import json
-import requests
+import re
 import urllib
 
-from bgmi.script import ScriptBase, parse_episode
-from bgmi.utils import print_error
-from bgmi.config import IS_PYTHON3
+import requests
 
+from bgmi.config import IS_PYTHON3
+from bgmi.script import ScriptBase
+from bgmi.utils import print_error, parse_episode
 
 if IS_PYTHON3:
     unquote = urllib.parse.unquote
